feat: redirect signed-in users from / to /predictions
The login form already redirects to /predictions after a successful
sign-in (when no ?next= is set), but signed-in users hitting the home
page (via the Header 'Home' link, the logo, or typing the URL) would
still see the marketing landing.

Refactor /app/page.tsx into a small server component that does an
auth check and redirects to /predictions when a session exists. The
existing client landing moves verbatim to HomePageContent.tsx and
remains the public face for signed-out visitors.
